# Optimise shops page: Enable injected enterprise data to be scoped to specific enterprise ids