--- a/lib/open_food_network/enterprise_injection_data.rb
+++ b/lib/open_food_network/enterprise_injection_data.rb
@@ -22,8 +22,10 @@ module OpenFoodNetwork
     end
 
     def shipping_method_services
-      @shipping_method_services ||= CacheService.cached_data_by_class("shipping_method_services",
-                                                                      Spree::ShippingMethod) do
+      @shipping_method_services ||= CacheService.cached_data_by_class(
+        "shipping_method_services_#{@enterprise_ids.hash}",
+        Spree::ShippingMethod
+      ) do
         # This result relies on a simple join with DistributorShippingMethod.
         # Updated DistributorShippingMethod records touch their associated Spree::ShippingMethod.
         Spree::ShippingMethod.services(@enterprise_ids)
@@ -31,7 +33,10 @@ module OpenFoodNetwork
     end
 
     def supplied_taxons
-      @supplied_taxons ||= CacheService.cached_data_by_class("supplied_taxons", Spree::Taxon) do
+      @supplied_taxons ||= CacheService.cached_data_by_class(
+        "supplied_taxons_#{@enterprise_ids.hash}",
+        Spree::Taxon
+      ) do
         # This result relies on a join with associated supplied products, through the
         # class Classification which maps the relationship. Classification records touch
         # their associated Spree::Taxon when updated. A Spree::Product's primary_taxon