--- a/spec/lib/open_food_network/enterprise_injection_data_spec.rb
+++ b/spec/lib/open_food_network/enterprise_injection_data_spec.rb
@@ -18,7 +18,7 @@ RSpec.describe OpenFoodNetwork::EnterpriseInjectionData do
   let!(:closed_oc) { create :closed_order_cycle, coordinator: enterprise4 }
 
   context "when scoped to specific enterprises" do
-    let(:subject) {
+    subject {
       described_class.new([enterprise1.id, enterprise2.id])
     }