#### What? Why?

When visiting https://openfoodnetwork.org.uk/shops, the page load is very slow. Given that this page is linked to from the
CTA above the fold on https://openfoodnetwork.org.uk/, it seems ideal that should load quickly.

It appears that part of the problem is that the `InjectionHelper` (via `OpenFoodNetwork::EnterpriseInjectionData`) fetches  information about _all_ enterprises (even ones which don't need to be serialised).

In some cases there could be a lot of rows pulled from the DB into memory, especially in the cases where there are joins (eg. `Spree::Taxon.supplied_taxons` pulls in all taxons of all products for all enterprises) - and then sometimes there are application-side operations happening on them too (eg. looping or creating hashes).

I opted to make this scoping optional where possible to try to ensure that they are non-breaking for other areas

I did a bit of benchmarking locally using
```
ab -n 20 -c 1 -H 'Accept: text/html'  -H 'Cookie: ...' 'http://localhost:3000/shops#/'
```
for this branch against master, and these were the results:

<details>
<summary>master (1940ms page load average):</summary>

```
This is ApacheBench, Version 2.3 <$Revision: 1903618 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient).....done


Server Software:        
Server Hostname:        localhost
Server Port:            3000

Document Path:          /shops#/
Document Length:        44150 bytes

Concurrency Level:      1
Time taken for tests:   38.817 seconds
Complete requests:      20
Failed requests:        19
   (Connect: 0, Receive: 0, Length: 19, Exceptions: 0)
Total transferred:      922991 bytes
HTML transferred:       884035 bytes
Requests per second:    0.52 [#/sec] (mean)
Time per request:       1940.863 [ms] (mean)
Time per request:       1940.863 [ms] (mean, across all concurrent requests)
Transfer rate:          23.22 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:  1697 1941 218.5   1815    2531
Waiting:     1696 1941 218.5   1815    2531
Total:       1697 1941 218.5   1815    2532

Percentage of the requests served within a certain time (ms)
  50%   1815
  66%   2015
  75%   2095
  80%   2132
  90%   2259
  95%   2532
  98%   2532
  99%   2532
 100%   2532 (longest request)
```
</details>

<details>
<summary>this branch (686ms page load average):</summary>

```
This is ApacheBench, Version 2.3 <$Revision: 1903618 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient).....done


Server Software:        
Server Hostname:        localhost
Server Port:            3000

Document Path:          /shops#/
Document Length:        43823 bytes

Concurrency Level:      1
Time taken for tests:   13.730 seconds
Complete requests:      20
Failed requests:        19
   (Connect: 0, Receive: 0, Length: 19, Exceptions: 0)
Total transferred:      914886 bytes
HTML transferred:       879896 bytes
Requests per second:    1.46 [#/sec] (mean)
Time per request:       686.478 [ms] (mean)
Time per request:       686.478 [ms] (mean, across all concurrent requests)
Transfer rate:          65.07 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.2      0       1
Processing:   543  686  91.4    671     955
Waiting:      543  686  91.4    670     955
Total:        543  686  91.4    671     956

Percentage of the requests served within a certain time (ms)
  50%    671
  66%    684
  75%    745
  80%    757
  90%    814
  95%    956
  98%    956
  99%    956
 100%    956 (longest request)
```
</details>

#### What should we test?
- The /shops page is the main one
- The groups page (eg /groups/1#/map) - from `inject_group_enterprises`
- The order show or edit page (eg /orders/R888114387) - from `inject_enterprise_and_relatives`

The other 2 pages incidentally also seem to be quicker for me, by a similar amount

#### Release notes

Changelog Category (reviewers may add a label for the release notes):

- [x] Technical changes only

#### Dependencies


#### Documentation updates