@johansenja 2024-08-13
Thanks for the review. It's useful to know about the rspec matchers, because I have a few other changes in mind which could benefit from them. With these changes, I don't think I'll be chaning the _number_ of queries, though there are a few areas which might (see https://github.com/openfoodfoundation/openfoodnetwork/pull/12724 - there are definitely some n+2s in there but I think I have a better solution for it)