@johansenja 2024-08-19
I am not sure about the new test failure now - I was wondering if it could be a flake, but it seems suspcious that it is in the order cycles area, though I'm not yet sure how my changes here would affect it. I'll take another look in due course