@drummer83 2024-08-19
Ja, I guess they are flaky. It's all green now.