@drummer83 2024-08-22
Hi @johansenja,
Thank you so much for working on the performance! It's so important, but often neglected unfortunately.
I staged your PR to our staging server and couldn't spot any issues. It's hard to tell if there is a significant improvement on the page loading time. I don't have the skills to use any tools for that.
However, I didn't find any problems, so this PR is ready to being merged.

For now I will move it to Ready To Go, to double check if we need another approval after you fixed the conflict.
@openfoodfoundation/core-devs Could you please check if another approval is required? Otherwise please go ahead and merge.

Thanks again! :partying_face: